✨ feat: Improve callback handling in create_autocmd()
This commit improves the callback handling in the `create_autocmd()` function by separating the callback from the other options and assigning it to a local variable. If the callback is a Lua function, it is assigned to `_autocmd[id]` and called via `require("artemis")._autocmd["' .. id .. '"]()`; if it is a Vimscript function, it is called via `vim.fn[callback](opts.arg)`. This simplifies the code and allows for better testing of the individual callbacks. Additionally, some minor formatting changes have been made to the file.